UPDATE/BUGFIX [23/12]
- correction prédiction : erreur stupide dans la lecture des arguments dans le main

- Dockerfile ajouté

- Début du re lancement des benchmarks et évaluations

+++ include/Operators/MAE.hpp
@@ -28,7 +28,7 @@ public:
   float compute(int position, const DataSet &data,
                 std::vector<int> index) const override;
 
-  static float apply(const std::vector<float> &exact,
+  static double apply(const std::vector<float> &exact,
                      const std::vector<float> &prediction);
 };
