+++ src/Operators/MAE.cpp
@@ -95,10 +95,10 @@ Parameters : exact results, prediction results
 Inputs     : const vector<float>, const vector<float>
 Outputs    : float
 */
-float MAE::apply(const std::vector<float> &exact,
+double MAE::apply(const std::vector<float> &exact,
                  const std::vector<float> &prediction) {
 
-  float res = 0;
+  double res = 0;
   float size = prediction.size();
   for (unsigned long int i = 0; i < size; ++i) {
     res += std::abs(exact[i] - prediction[i]);
